Delete old version of SearchField component, rename new version
